assessment component organized in folder
