+++ src/components/AssessmentComponents/WeightageList/scss/WeightageList.css
