+++ src/components/AssessmentComponents/WeightageList/scss/WeightageList.scss
